refactor(handlers): extract handlerUtil
